requirements.txt
annotated-types==0.7.0
anyio==4.11.0
art==6.5
astroid==4.0.2
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
dill==0.4.0
fonttools==4.61.1
gitingest==0.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
isort==7.0.0
kiwisolver==1.4.9
loguru==0.7.3
markdown-it-py==4.0.0
mccabe==0.7.0
mdurl==0.1.2
numpy==2.3.4
packaging==25.0
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.1
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pylint==4.0.4
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-toon==0.1.3
pytz==2025.2
regex==2025.11.3
requests==2.32.5
six==1.17.0
sniffio==1.3.1
starlette==0.50.0
thespian==4.0.1
tiktoken==0.12.0
tomlkit==0.13.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
